api/image-api/pyproject.toml
[tool.poetry]
name = "image_api"
version = "v0.1.0"
description = "A text-to-image and image-to-image server compatible with the OpenAI API"
authors = ["GalaxyAIOS Authors <contact@galaxyaios.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
torch = ">=2.0.0"
torchvision = "*"
diffusers = ">=0.31.0"
transformers = ">=4.47.0"
accelerate = ">=1.0.0"
safetensors = ">=0.4.0"
huggingface-hub = ">=0.25.0"
modelscope = ">=1.20.0"
fastapi = ">=0.115.0"
uvicorn = ">=0.32.0"
python-multipart = ">=0.0.17"
pydantic = ">=2.0.0"
pillow = ">=10.0.0"
httpx = ">=0.27.0"
aiofiles = ">=23.0.0"
numpy = ">=1.24.0"
sentencepiece = "*"
protobuf = "*"
bitsandbytes = { version = ">=0.44.0", markers = "platform_system == 'Linux'" }

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-asyncio = ">=0.21.0"
httpx = ">=0.27.0"
flake8 = ">=7.0.0"
black = ">=24.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
image-api = "image_api.main:main"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
